Project BlackHawk 0.1.0 Bugfix: Captcha and Install
In the Captcha class, we had a bug that prevented correct captcha values
from being validated as correct. This has been fixed.
In the setup, towards the end of the install, we delete the /setup/
folder. This has now become a mandatory function (no longer optional).
In doing this, the verbage of the Continue button has changed. The
delete folder function, for development purposes, remains disabled.